Add dynamic draw and position indicator
Add a red box under the cursor to inidcate the current position of the cursor, allowing you to more accurately draw shapes and spacing.

Added the check for all shapes currently on screen being drawn, to prevent a large number of off screen shapes from being drawn every frame which do not need to be.

Included prints in the console for time drawing the board (and number of shapes drawn) as well as time taken to calculate the current board.