Prebuilt Ship Stamps
Added the ability to place pre-built ships into the map, with the name of the current stamp being displayed in text at the bottom of the screen. The current stamp is changed using the mouse wheel, with single being the initial state, then alphabetically past that. A better method of chosing and distinguising the purposes for the stamps should in future be created.